Fix annotations/highlighing of macro arguments.
One problem is that libclang's clang_annotate sometimes yields an uspecific
token https://llvm.org/bugs/show_bug.cgi?id=12972. In macros this problem is
worse: If one does not specify CXTranslationUnit_DetailedPreprocessingRecord,
one only gets the AST down to CompoundStmt which is useless. If one does specify
the option, clang_getCursor() on a location inside the macro expansion will
yield the MacroExpansion cursor which is detached from the main AST so the
interesting cursor is not a child.

The first attempt to fix this by using clang_visitChildren over the unspecific
annotation cursor yielded by clang_annotate failed because it was not only
unspecific bad wrong (Ex.: assert(c) expands to
(c ? (void)0 : __assert_fail(...) but all tokens inside the expansion where
annotated with the CallExpr of __assert_fail.

The second attempt was to work around these bugs by using
clang_getCursorSemanticParent or clang_getCursorLexicalParent on the wrong
annotation cursor as long as the start location of the cursor is equal to the
start location of the macro expansion and visit children from there. But as it
turns out, these functions seem to yield bogus results (e.g. the previous
sibling of the actual parent or an IvalidFile cursor) on the annotation cursors.

Thus the final fix is to iterate over the AST of the whole translation unit and
attempt to annotate all bad tokens.

If your asserts still aren't getting annotated you are probably compiling in
Release mode (i.e. with NDEBUG defined).