Handle C in fileUniqueName.
For types in C, we can have a function/variable/typedef S, a "struct S",
an "enum S" and a "union S" all at the same time. Thus we need to
encode that in the name.
Also, since functions cannot be overloaded in C, simply use the name
there.

To determine if a translation unit is in C, we check if any Cursor in
the TU's AST is non-C (and non-invalid) because clang_getCursorLanguage
sometimes eratically returns C for a cursor in a C++ TU.