Set fileUniqueName only on destination markup.
This fixes icorrectly generated links to some compiler-generated
functions, e.g. in synth's source code to
CXString.h.html#operator=:CXString&&.
The links are still generated but at least point to the line number of
non-existent function which happens to be the "containing" truct/class.
As a side-effect, this is a simplification of the code too.